Send heading as int
The magnetometer and calculations are not precise enough to warrant a float being sent. This may also fix some problems with the messages being lagged behind, as it saves some bytes in the packet.

+++ src/main/java/tech/mistermel/edisoncontrol/navigation/NavigationHandler.java
@@ -69,7 +69,7 @@ public class NavigationHandler {
 	}
 	
 	private void tick() {
-		EdisonControl.getInstance().getWebHandler().sendPosition(x, y, heading);
+		EdisonControl.getInstance().getWebHandler().sendPosition(x, y, (int) heading);
 		
 		if(!EdisonControl.getInstance().getDWMSerialInterface().isCommunicationWorking()) {
 			logger.warn("DWM serial is not working! Exiting navigation mode.");
